test function moved into existing file.
