+++ fsspec/implementations/tests/test_sftp.py
@@ -104,6 +104,19 @@ def test_transaction(ssh):
     ]
 
 
+def test_mkdir_create_parent(ssh):
+    f = fsspec.get_filesystem_class("sftp")(**ssh)
+
+    with pytest.raises(FileNotFoundError):
+        f.mkdir("/a/b/c")
+
+    f.mkdir("/a/b/c", create_parents=True)
+    assert f.exists("/a/b/c")
+
+    with pytest.raises(FileExistsError, match="/a/b/c"):
+        f.mkdir("/a/b/c")
+
+
 def test_makedirs_exist_ok(ssh):
     f = fsspec.get_filesystem_class("sftp")(**ssh)
 
@@ -113,3 +126,4 @@ def test_makedirs_exist_ok(ssh):
         f.makedirs("/a/b/c", exist_ok=False)
 
     f.makedirs("/a/b/c", exist_ok=True)
+
